requirements.txt
flask==3.1.3
requests==2.32.5
beautifulsoup4==4.14.3
Pillow==11.1.0
functions-framework==3.8.2
google-cloud-firestore==2.20.1
